pyproject.toml
[tool.poetry]
name = "sep-cogs"
version = "0.1.0"
description = "(Possibly) useful utilities and tools for your Discord server's Red-DiscordBot (V3)."
authors = ["Seputaes <seputaes@sep.gg>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = ">=3.6.2,<3.8"
aiohttp="3.4.4"
aiohttp-json-rpc="0.11.2"
appdirs="1.4.3"
async-timeout="3.0.1"
attrs="18.2.0"
chardet="3.0.4"
colorama="0.4.1"
distro="1.3.0"
dnspython="1.16.0"
fuzzywuzzy="0.17.0"
idna="2.8"
idna-ssl="1.1.0"
motor="2.0.0"
multidict="4.5.2"
pymongo="3.7.2"
python-Levenshtein="0.12.0"
PyYAML="3.13"
raven="6.10.0"
raven-aiohttp="0.7.0"
Red-Lavalink="0.1.2"
schema="0.6.8"
websockets="6.0"
yarl="1.3.0"
"discord.py" = { git = "https://github.com/Rapptz/discord.py", rev = "7f4c57dd5ad20b7fa10aea485f674a4bc24b9547" }
red-discordbot = { git = "https://github.com/Cog-Creators/Red-DiscordBot.git", branch = "V3/develop", extras = ["voice", "mongo"] }

[tool.poetry.dev-dependencies]
red-discordbot = { git = "https://github.com/Seputaes/Red-DiscordBot.git", branch = "V3/develop" }
tox = "^3.7"
toml = "*"
black = "18.9b0"
pytest = "*"
pytest-asyncio = "*"
pytz = "^2018.9"
recurrent = "^0.2.5"
timeago = "^1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
